requirements.txt
Pillow
pywin32
bettercam
configparser
opencv-python
numpy
colorama
hidapi